rework: implement already signed up / preparing wallet /  success screen without adding new routes
