+++ src/SignupRouter.js
@@ -11,7 +11,7 @@ import SigninInfo from './components/signin/SigninInfo'
 import Blurred from './components/common/view/Blurred'
 import Welcome from './components/auth/login/WelcomeScreen'
 import InviteWelcome from './components/inviteWelcome/InviteWelcome'
-import AccountAlreadyExists from './components/auth/AccountAlreadyExists'
+import { AuthContextProvider } from './components/auth/context/AuthContext'
 
 const log = logger.child({ from: 'SignupRouter' })
 
@@ -20,7 +20,6 @@ const generateRouter = () => {
   const { enableSelfCustody } = Config
 
   const routes = {
-    AccountAlreadyExists,
     Welcome,
     InviteWelcome,
     Signup,
@@ -45,7 +44,9 @@ const Router = () => {
   return (
     <>
       <Blurred whenDialog>
-        <RouterWrapper onNavigationStateChange={navigationStateHandler} />
+        <AuthContextProvider>
+          <RouterWrapper onNavigationStateChange={navigationStateHandler} />
+        </AuthContextProvider>
       </Blurred>
     </>
   )
