+++ src/components/auth/login/SignUpScreen.js
@@ -4,7 +4,7 @@ import { Platform, View } from 'react-native'
 import Wrapper from '../../common/layout/Wrapper'
 import Text from '../../common/view/Text'
 import NavBar from '../../appNavigation/NavBar'
-import AuthProgressBar from '../AuthProgressBar'
+import AuthProgressBar from '../components/AuthProgressBar'
 import { withStyles } from '../../../lib/styles'
 import { theme as mainTheme } from '../../theme/styles'
 
@@ -20,6 +20,7 @@ import googleBtnIcon from '../../../assets/Auth/btn-google.svg'
 import facebookBtnIcon from '../../../assets/Auth/btn-facebook.svg'
 import Config from '../../../config/config'
 import logger from '../../../lib/logger/js-logger'
+import AuthStateWrapper from '../components/AuthStateWrapper'
 import { LoginButton } from './LoginButton'
 import Recaptcha from './Recaptcha'
 
@@ -99,132 +100,134 @@ const SignupScreen = ({ screenProps, styles, handleLoginMethod, sdkInitialized, 
     <Wrapper backgroundColor="#fff" style={styles.mainWrapper}>
       <NavBar logo />
       <AuthProgressBar step={1} />
-      <Section.Stack style={{ flex: 1, justifyContent: 'center' }}>
-        <Section.Stack style={{ flex: 1, maxHeight: 640 }}>
-          <Section.Stack style={{ flexGrow: 0 }}>
-            <Text
-              color={'primary'}
-              fontSize={getDesignRelativeHeight(12)}
-              lineHeight={getDesignRelativeHeight(21)}
-              letterSpacing={0.26}
-              fontFamily="Roboto"
-              fontWeight="bold"
-              textTransform="uppercase"
-            >
-              Choose Authentication Method
-            </Text>
-            <Text
-              color={'darkIndigo'}
-              fontSize={getDesignRelativeHeight(26)}
-              lineHeight={getDesignRelativeHeight(34)}
-              letterSpacing={0.26}
-              fontFamily="Roboto"
-              fontWeight="bold"
-              style={{ marginTop: getDesignRelativeHeight(15) }}
-            >
-              Simply Claim G$ Daily
-            </Text>
-            <Text
-              color={'darkIndigo'}
-              fontSize={getDesignRelativeHeight(18)}
-              lineHeight={getDesignRelativeHeight(23)}
-              letterSpacing={0.26}
-              fontFamily="Roboto"
-              style={{ marginTop: getDesignRelativeHeight(5) }}
-            >
-              {`Your security is important to us, so we\n need to verify you’re a real person.\n Continuing using Facebook or Google\n would make that easier.`}
-            </Text>
-          </Section.Stack>
-          <Section.Stack style={styles.bottomContainer}>
-            <View style={{ width: '100%' }}>
-              <LoginButton
-                style={[styles.buttonLayout, { backgroundColor: mainTheme.colors.googleRed }]}
-                onPress={_google}
-                disabled={!sdkInitialized}
-                testID="login_with_google"
-                icon={googleBtnIcon}
+      <AuthStateWrapper>
+        <Section.Stack style={{ flex: 1, justifyContent: 'center' }}>
+          <Section.Stack style={{ flex: 1, maxHeight: 640 }}>
+            <Section.Stack style={{ flexGrow: 0 }}>
+              <Text
+                color={'primary'}
+                fontSize={getDesignRelativeHeight(12)}
+                lineHeight={getDesignRelativeHeight(21)}
+                letterSpacing={0.26}
+                fontFamily="Roboto"
+                fontWeight="bold"
+                textTransform="uppercase"
               >
-                {`${buttonPrefix} Google`}
-              </LoginButton>
-              <LoginButton
-                style={[
-                  styles.buttonLayout,
-                  styles.buttonsMargin,
-                  {
-                    backgroundColor: mainTheme.colors.facebookBlue,
-                  },
-                ]}
-                onPress={_facebook}
-                disabled={!sdkInitialized}
-                testID="login_with_facebook"
-                icon={facebookBtnIcon}
-                iconProps={{ viewBox: '0 0 11 22' }}
+                Choose Authentication Method
+              </Text>
+              <Text
+                color={'darkIndigo'}
+                fontSize={getDesignRelativeHeight(26)}
+                lineHeight={getDesignRelativeHeight(34)}
+                letterSpacing={0.26}
+                fontFamily="Roboto"
+                fontWeight="bold"
+                style={{ marginTop: getDesignRelativeHeight(15) }}
               >
-                {`${buttonPrefix} Facebook`}
-              </LoginButton>
-
-              <Recaptcha ref={reCaptchaRef} onSuccess={onRecaptchaSuccess} onFailure={onRecaptchaFailed}>
+                Simply Claim G$ Daily
+              </Text>
+              <Text
+                color={'darkIndigo'}
+                fontSize={getDesignRelativeHeight(18)}
+                lineHeight={getDesignRelativeHeight(23)}
+                letterSpacing={0.26}
+                fontFamily="Roboto"
+                style={{ marginTop: getDesignRelativeHeight(5) }}
+              >
+                {`Your security is important to us, so we\n need to verify you’re a real person.\n Continuing using Facebook or Google\n would make that easier.`}
+              </Text>
+            </Section.Stack>
+            <Section.Stack style={styles.bottomContainer}>
+              <View style={{ width: '100%' }}>
+                <LoginButton
+                  style={[styles.buttonLayout, { backgroundColor: mainTheme.colors.googleRed }]}
+                  onPress={_google}
+                  disabled={!sdkInitialized}
+                  testID="login_with_google"
+                  icon={googleBtnIcon}
+                >
+                  {`${buttonPrefix} Google`}
+                </LoginButton>
                 <LoginButton
                   style={[
                     styles.buttonLayout,
                     styles.buttonsMargin,
                     {
-                      backgroundColor: mainTheme.colors.white,
-                      borderWidth: 1,
-                      borderColor: '#E9ECFF',
+                      backgroundColor: mainTheme.colors.facebookBlue,
                     },
                   ]}
-                  onPress={_mobile}
+                  onPress={_facebook}
                   disabled={!sdkInitialized}
-                  textColor="#8499BB"
-                  testID="login_with_auth0"
+                  testID="login_with_facebook"
+                  icon={facebookBtnIcon}
+                  iconProps={{ viewBox: '0 0 11 22' }}
                 >
-                  {`${buttonPrefix} Passwordless`}
+                  {`${buttonPrefix} Facebook`}
                 </LoginButton>
-              </Recaptcha>
-            </View>
-            <Section.Stack style={styles.textButtonContainer}>
-              {Config.enableSelfCustody && (
-                <View>
-                  <CustomButton
-                    compact
-                    mode={'text'}
-                    color={mainTheme.colors.darkGray}
-                    textStyle={{
-                      textDecorationLine: 'underline',
-                      fontSize: 14,
-                      fontWeight: 'bold',
-                      lineHeight: 16,
-                      letterSpacing: 0.14,
-                    }}
-                    onPress={_selfCustody}
-                    style={styles.textButton}
+
+                <Recaptcha ref={reCaptchaRef} onSuccess={onRecaptchaSuccess} onFailure={onRecaptchaFailed}>
+                  <LoginButton
+                    style={[
+                      styles.buttonLayout,
+                      styles.buttonsMargin,
+                      {
+                        backgroundColor: mainTheme.colors.white,
+                        borderWidth: 1,
+                        borderColor: '#E9ECFF',
+                      },
+                    ]}
+                    onPress={_mobile}
+                    disabled={!sdkInitialized}
+                    textColor="#8499BB"
+                    testID="login_with_auth0"
                   >
-                    Self Custody SignUp
-                  </CustomButton>
-                  <CustomButton
-                    compact
-                    mode={'text'}
-                    color={mainTheme.colors.darkGray}
-                    textStyle={{
-                      textDecorationLine: 'underline',
-                      fontSize: 14,
-                      fontWeight: 'bold',
-                      lineHeight: 16,
-                      letterSpacing: 0.14,
-                    }}
-                    onPress={_selfCustodyLogin}
-                    style={styles.textButton}
-                  >
-                    Self Custody Login
-                  </CustomButton>
-                </View>
-              )}
+                    {`${buttonPrefix} Passwordless`}
+                  </LoginButton>
+                </Recaptcha>
+              </View>
+              <Section.Stack style={styles.textButtonContainer}>
+                {Config.enableSelfCustody && (
+                  <View>
+                    <CustomButton
+                      compact
+                      mode={'text'}
+                      color={mainTheme.colors.darkGray}
+                      textStyle={{
+                        textDecorationLine: 'underline',
+                        fontSize: 14,
+                        fontWeight: 'bold',
+                        lineHeight: 16,
+                        letterSpacing: 0.14,
+                      }}
+                      onPress={_selfCustody}
+                      style={styles.textButton}
+                    >
+                      Self Custody SignUp
+                    </CustomButton>
+                    <CustomButton
+                      compact
+                      mode={'text'}
+                      color={mainTheme.colors.darkGray}
+                      textStyle={{
+                        textDecorationLine: 'underline',
+                        fontSize: 14,
+                        fontWeight: 'bold',
+                        lineHeight: 16,
+                        letterSpacing: 0.14,
+                      }}
+                      onPress={_selfCustodyLogin}
+                      style={styles.textButton}
+                    >
+                      Self Custody Login
+                    </CustomButton>
+                  </View>
+                )}
+              </Section.Stack>
+              <SignupText />
             </Section.Stack>
-            <SignupText />
           </Section.Stack>
         </Section.Stack>
-      </Section.Stack>
+      </AuthStateWrapper>
     </Wrapper>
   )
 }
